Creating a more generic rule for HTML
